backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic-settings
python-multipart
pandas
numpy
scipy
openpyxl
neurokit2
opencv-python-headless
pillow
matplotlib
openai
torch
torchvision
torchaudio
albumentations
opencv-python-headless
scikit-image
scikit-learn
tqdm
pillow
matplotlib
seaborn